+++ database/migrations/2026_03_15_105126_create_companies_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('name', length: 255);
             $table->string('registration_no', length: 20);
-            $table->float('tax_value', precision: 2)->default(20.00);
+            $table->decimal('tax_value', 4, 2)->default(20.00);
             $table->string('invoice_prefix')->nullable(false);
             $table->timestamps();
             $table->softDeletes();
